NAPVAL Build fixes
Connections to stan/nats servers not consistent. NAPVAL as separate
build now without dependencies on sms or 2 instances of stan.
Multiple other small cleanups and refactors.
