--- app/napval/stopnias.sh
@@ -1,3 +1,0 @@
-kill -9 `cut -f 1 nias.pid`
-rm nias.pid
-echo "NIAS services shut down"
