+++ bin/gonias.bat
@@ -5,7 +5,7 @@ rem gnatsd MUST be the first program launched
 @echo off
 
 rem Run the NIAS services. Add to the BOTTOM of this list
-start nats-streaming-server -sc napval_nss.cfg
+start nats-streaming-server -p 4223 -sc napval_nss.cfg
 start napval
 
 rem Run the web client (launch browser here)
