+++ build_napval.sh
@@ -19,7 +19,7 @@ do_build() {
 	go get 
 	GOOS="$GOOS" GOARCH="$GOARCH" go build -ldflags="$LDFLAGS" -o $OUTPUT/$HARNESS
 	cd ..
-	rsync -a ../test_data napval/nias8help.pdf napval/napval.toml napval/nss.cfg napval/public napval/schemas napval/schoolslist napval/templates $OUTPUT/
+	rsync -a napval/nias8help.pdf napval/napval.toml napval/napval_nss.cfg napval/public napval/schemas napval/schoolslist napval/templates $OUTPUT/
 }
 
 do_shells() {
@@ -47,7 +47,7 @@ do_goupx() {
 do_zip() {
 	cd $OUTPUT
 	cd ..
-	zip -qr ../$ZIP go-nias8
+	zip -qr ../$ZIP napval
 	cd $CWD
 }
 
@@ -57,10 +57,10 @@ build_mac64() {
 	GOOS=darwin
 	GOARCH=amd64
 	LDFLAGS="-s -w"
-	OUTPUT=$CWD/build/Mac/go-nias8
+	OUTPUT=$CWD/build/Mac/napval
 	GNATS=nats-streaming-server
 	HARNESS=napval
-	ZIP=go-nias-Mac.zip
+	ZIP=nias-napval-Mac.zip
 	do_build
 	#do_upx
 	do_shells
@@ -75,10 +75,10 @@ build_windows64() {
 	GOOS=windows
 	GOARCH=amd64
 	LDFLAGS="-s -w"
-	OUTPUT=$CWD/build/Win64/go-nias8
+	OUTPUT=$CWD/build/Win64/napval
 	GNATS=nats-streaming-server.exe
 	HARNESS=napval.exe
-	ZIP=go-nias-Win64.zip
+	ZIP=nias-napval-Win64.zip
 	do_build
 	#do_upx
 	do_bats
@@ -92,10 +92,10 @@ build_windows32() {
 	GOOS=windows
 	GOARCH=386
 	LDFLAGS="-s -w"
-	OUTPUT=$CWD/build/Win32/go-nias8
+	OUTPUT=$CWD/build/Win32/napval
 	GNATS=nats-streaming-server.exe
 	HARNESS=napval.exe
-	ZIP=go-nias-Win32.zip
+	ZIP=nias-napval-Win32.zip
 	do_build
 	#do_upx
 	do_bats
@@ -109,10 +109,10 @@ build_linux64() {
 	GOOS=linux
 	GOARCH=amd64
 	LDFLAGS="-s -w"
-	OUTPUT=$CWD/build/Linux64/go-nias8
+	OUTPUT=$CWD/build/Linux64/napval
 	GNATS=nats-streaming-server
 	HARNESS=napval
-	ZIP=go-nias-Linux64.zip
+	ZIP=nias-napval-Linux64.zip
 	do_build
 	#do_goupx
 	do_shells
@@ -126,10 +126,10 @@ build_linux32() {
 	GOOS=linux
 	GOARCH=386
 	LDFLAGS="-s -w"
-	OUTPUT=$CWD/build/Linux32/go-nias8
+	OUTPUT=$CWD/build/Linux32/napval
 	GNATS=nats-streaming-server
 	HARNESS=napval
-	ZIP=go-nias-Linux32.zip
+	ZIP=nias-napval-Linux32.zip
 	do_build
 	#do_goupx
 	do_shells
